+++ lruReplacementPolicy.c
@@ -13,7 +13,7 @@
 /**
  * This will delete the least recently used node with the specified key
  */
-rbtree_node* replace(rbtree_node *root, int pid, int vpn, unsigned long timeCreated) {
+rbtree_node* replace(rbtree_node *root, unsigned long pid, unsigned long vpn, unsigned long timeCreated) {
     rbtree_node *temp_node = searchForLRU(root);
     root = rbtree_delete_node(root, temp_node->key);
     root = rbtree_insert(root, vpn, pid, timeCreated, true);
